[breadboard-ui] Change node order and tidy a few pieces (#353)
Previously I had it set so that `graphstart` and `graphend` would wrap
the `nodestart` and `nodeend`, which I did with a bit of wrangling.

Since the semantics of it are the other way around – that we should see
`invoke` _before_ `graphstart` etc – I've made it so. This has required
a small workaround in the code, such that if there's an existing node
with the ID of the to-be-inserted one, I update the to-be-inserted one
to be a child of the existing one and I update its ID.

For example, if we have `nodestart` `[2]` and then `graphstart` `[2]`, I
update `graphstart`'s path to `[2,0]` and make it a child of `nodestart`
`[2]`.