[breadboard-ui] Make inputs based on state (#357)
This does some pretty sizeable changes to the UI around inputs. In
particular it attempts to render inputs & secrets based on whichever
messages have come in. This should pave the way for us to traverse the
messages we've received and for the UI to update accordingly.

I'll land this as-is, but there may be wrinkles that need to be ironed
out!