[breadboard-ui] Fix for inputs failing to ask for data (#489)
In the `input-list` we work through all the messages we've received so
far. We do this so that we create the UI entirely from the messages
received and we don't base it on any internal state. So that we can
stack newer inputs first we work through the list backwards. When we
find an input we then track forwards again to see if there are any
subsequent messages we received that tell us what the input value should
be. So, for example, an input in the middle of the messages that has
already been answered by the user will have a subsequent message with
its value set. We use that to populate the field and make it readonly.

There was a bug in the code where we do that, and it wasn't traversing
correctly. As such it would mistakenly switch off inputs we'd already
seen because it would find an "answer" message for the older input of
the same ID.

This PR fixes all that.