[breadboard-extension] Add user setting for board locations (#492)
To try and smooth things out a bit for choosing sub boards I've added a
user configuration option. When a user runs the extension for the first
time (or if they clear the setting), the extension will ask them where
the JSON boards are located. It then uses this path whenever it
encounters an input value whose default ends in `.json`.

It would be good to figure out a neater way to encode the location of
sub boards, though, as this approach is somewhat brittle and will only
work if the user is asked for a particular board location as part of the
outer board's execution.