[breadboard-web/-ui] Enable unstarted, blank boards (#666)
I'm exploring the idea of creating a board in the debugger / playground
UI. For that to work, however, we need to be able to boot the Build view
without a URL set. This PR shouldn't affect the current approach, which
is to say:

1. If no board URL is set, we will show the List view of boards.
2. If a board is clicked in the list we will switch to the Build view
and autorun the board.
3. If the debugger is loaded with a board URL set, we will switch to the
Build view and autorun the board.

But now there is an additional option:

4. If the user clicks on "New board" in the List view (or deeplinks to
`?mode=build&visualizer=editor`) we load the Build view with the editor
UI (to be built) and the debugger does not attempt to run the board.