[breadboard-ui] Add "use defaults" button (#858)
I noticed that when we drag a new node on we automatically populate the
properties with the examples/defaults, and this mistakenly gives the
impression that the node is configured. This PR changes it so that you
have to click the "use defaults" button for the form to be populated
with the info. The additional benefit of this is that you can reset the
configuration at any time with this, too.