[build] Compatibility with runner and inspector (#1176)
Nodes defined with the `@breadboard-ai/build` package are now compatible
with the Breadboard runner and inspector API.

Just required a small tweak to the way those APIs check whether
something looks like a node type. Previously there were short-circuits
for when something was a function, but since `@breadboard-ai/build` node
definitions are themselves functions (for use with its `board` API), we
instead now just look for an `invoke` or `describe` property.

This also adds a new `NodeFactoryFromDefinition` type utility, which
generates a signature for a node type so that it doesn't need to be
written out manually when building kits.

Also added integration tests.