[build] [core-kit] [template-kit] Be more explicit & verbose with generated schemas (#1533)
- For `unknown`, we previously emitted `{}`, but now we emit `{type:
["array", "boolean", "null", "number", "object", "string"]}`. These are
technically equivalent according to JSON schema, but since there is some
other Breadboard code that treats `{}` as `string`, doing the verbose
thing is a bit more compatible.

- Updated some `core-kit` and `template-kit` nodes to use `unknown`
instead of `object`. In JSON Schema, `object` means a JSON object like
`{foo}`, but not a basic type like `"foo"`. In these nodes, some
inputs/outputs can be any kind of JSON, so this is now the more correct
broader type.