[build] Add converge function for ports with multiple incoming edges (#1736)
Previously, nodes could only be instantiated with a single input,
meaning an input port could only have one incoming edge. Now using the
`converge` function, you can pass multiple edges.

```ts
def({ foo: converge(a, b) }
```

The alternative to this syntax was accepting arrays of input values.
However, I think it's useful to have a clearly distinctive signature to
remove any ambiguity between *an array of values intended to each be
separately wired* (the convergence case), from *an array of values that
is a single input for that port* (the case where a port accepts an
array).